# Back-end/src/main/java/com/projeto/api/domain/habits/HabitDTO.java
package com.projeto.api.domain.habits;

import jakarta.validation.constraints.NotBlank;

public record HabitDTO(@NotBlank String nomeHabito, Frequency frequencia, Period periodo, Completed completed) {
}